split for file length
