+++ src/routes/applet/dot_product/diagonal_parallelogram/+page.svelte
@@ -4,28 +4,15 @@
   import { Vector2, Vector3 } from 'three';
   import { GridType } from '$lib/d3-components/grids/GridTypes';
   import { Formula } from '$lib/utils/Formulas';
-  import Arc from '$lib/d3-components/Arc.svelte';
   import RightAngle from '$lib/d3-components/RightAngle.svelte';
   import { snapPointToLine } from '$lib/utils/MathLib';
+  import StaticImg from './staticImg.svelte';
 
   //[] todo consistant var naming -> camelcase or __
-  //[X] todo move labels
-  //[] todo make version of grid class with only snap function so orthogonality check works
-  //[X] todo new formula slot
-  //[] refactor length of this file
 
   const snap_distance = 0.5;
 
   let formulas: Formula[] = [];
-
-  //vars for static section
-  let o_static = new Vector2(-4, -1);
-
-  let w_static = new Vector2(1.5, 3);
-  let v_static = new Vector2(2, -1);
-
-  let vPlusW_static = v_static.clone().add(w_static);
-  let vMinusW_static = v_static.clone().sub(w_static);
 
   //vars for interactive section
   let origin = new Vector2(1.5, 0);
@@ -86,6 +73,8 @@
 </script>
 
 <Canvas2D gridType={GridType.None} {formulas}>
+  <!-- Imported static side -->
+  <StaticImg />
   <!-- Interactive side  of image -->
 
   <Draggable2D id="w+o" bind:position={w_plus_o} color={PrimeColor.darkGreen} />
@@ -159,96 +148,4 @@
     hideHead
   />
   <RightAngle {origin} vs={[v, w]} />
-
-  <!-- Static part of image -->
-
-  <!-- static Bases -->
-  <Vector2D
-    origin={o_static}
-    direction={v_static}
-    length={v_static.length()}
-    color={PrimeColor.cyan}
-    let:endPoint
-  >
-    <Latex2D
-      position={endPoint}
-      latex={'\\mathbf{v}'}
-      offset={new Vector2(-0.4, 0.5)}
-      color={PrimeColor.cyan}
-    />
-  </Vector2D>
-  <Vector2D
-    origin={o_static}
-    direction={w_static}
-    length={w_static.length()}
-    color={PrimeColor.darkGreen}
-    let:endPoint
-  >
-    <Latex2D
-      position={endPoint}
-      latex={'\\mathbf{w}'}
-      offset={new Vector2(-0.2, 0.2)}
-      color={PrimeColor.darkGreen}
-    />
-  </Vector2D>
-
-  <!-- static v+w -->
-  <Vector2D
-    origin={o_static}
-    direction={vPlusW_static}
-    length={vPlusW_static.length()}
-    color={PrimeColor.raspberry}
-    let:endPoint
-  >
-    <Latex2D
-      position={endPoint}
-      latex={'\\mathbf{v + w}'}
-      offset={new Vector2(0.2, 0.2)}
-      color={PrimeColor.raspberry}
-    />
-  </Vector2D>
-
-  <!-- static v-w -->
-  <Vector2D
-    origin={w_static.clone().add(o_static)}
-    direction={vMinusW_static}
-    length={vMinusW_static.length()}
-    color={PrimeColor.orange}
-    let:endPoint
-  >
-    <Latex2D
-      position={endPoint}
-      latex={'\\mathbf{v - w}'}
-      offset={new Vector2(-0.2, -0.2)}
-      color={PrimeColor.orange}
-    />
-  </Vector2D>
-
-  <!-- Static Helper lines -->
-  <Vector2D
-    origin={w_static.clone().add(o_static)}
-    direction={v_static}
-    length={v_static.length()}
-    color={PrimeColor.black}
-    isDashed
-    hideHead
-  />
-  <Vector2D
-    origin={v_static.clone().add(o_static)}
-    direction={w_static}
-    length={w_static.length()}
-    color={PrimeColor.black}
-    isDashed
-    hideHead
-  />
-
-  <RightAngle origin={o_static} vs={[v_static, w_static]} />
-
-  <!-- <div slot="formulas">
-      {#if isOrthogonal}
-        <LatexUI params={[v.x]} latex={' v + w = v - w'} />
-      {:else}
-        <LatexUI params={[w.y]} latex={' v + w \\neq v - w'} />
-      {/if}
-    </div> -->
 </Canvas2D>
